Split QwiicButton class into partial files
