+++ src/devices/QwiicButton/QwiicButton.cs
@@ -11,12 +11,10 @@ namespace Iot.Device.QwiicButton
     /// SparkFun Qwiic Button is an I2C based button with a built-in LED.
     /// Supported hardware version: 1.0.0
     /// </summary>
-    public class QwiicButton : IDisposable
+    public partial class QwiicButton : IDisposable
     {
         private const int DefaultAddress = 0x6F; // Default I2C address of the button
         private I2cBusAccess _i2cBus;
-
-        /*-------------------------------- Device Status ------------------------*/
 
         /// <summary>
         /// Initializes a new instance of the <see cref="QwiicButton"/> class.
@@ -52,7 +50,8 @@ namespace Iot.Device.QwiicButton
 
         /// <summary>
         /// Returns the firmware version of the button as a 16-bit integer.
-        /// The leftmost (high) byte is the major revision number, and the rightmost (low) byte is the minor revision number.
+        /// The leftmost (high) byte is the major revision number, and the rightmost (low) byte is
+        /// the minor revision number.
         /// </summary>
         public ushort GetFirmwareVersionAsInteger()
         {
@@ -92,302 +91,6 @@ namespace Iot.Device.QwiicButton
             Console.WriteLine("Could not set I2C address");
             return false;
         }
-
-        /*------------------------------ Button Status ---------------------- */
-
-        /// <summary>
-        /// Returns whether the button is pressed.
-        /// </summary>
-        public bool IsPressed()
-        {
-            var status = new StatusRegisterBitField(_i2cBus.ReadSingleRegister(Register.BUTTON_STATUS));
-            return status.IsPressed;
-        }
-
-        /// <summary>
-        /// Returns whether the button is clicked.
-        /// </summary>
-        public bool HasBeenClicked()
-        {
-            var status = new StatusRegisterBitField(_i2cBus.ReadSingleRegister(Register.BUTTON_STATUS));
-            return status.HasBeenClicked;
-        }
-
-        /// <summary>
-        /// Returns whether a new button status event has occurred.
-        /// </summary>
-        public bool EventAvailable()
-        {
-            var status = new StatusRegisterBitField(_i2cBus.ReadSingleRegister(Register.BUTTON_STATUS));
-            return status.EventAvailable;
-        }
-
-        /// <summary>
-        /// Sets <see cref="IsPressed"/>, <see cref="HasBeenClicked"/> and <see cref="EventAvailable"/> to false.
-        /// </summary>
-        public byte ClearEventBits()
-        {
-            var status = new StatusRegisterBitField(_i2cBus.ReadSingleRegister(Register.BUTTON_STATUS))
-            {
-                EventAvailable = false,
-                HasBeenClicked = false,
-                IsPressed = false
-            };
-            return _i2cBus.WriteSingleRegisterWithReadback(Register.BUTTON_STATUS, status.StatusRegisterValue);
-        }
-
-        /// <summary>
-        /// Returns the time in milliseconds that the button waits for the mechanical contacts to settle.
-        /// </summary>
-        public ushort GetDebounceTime()
-        {
-            return _i2cBus.ReadDoubleRegister(Register.BUTTON_DEBOUNCE_TIME);
-        }
-
-        /// <summary>
-        /// Sets the time in milliseconds that the button waits for the mechanical contacts to settle and checks if the register was set properly.
-        /// Returns 0 on success, 1 on register I2C write fail, and 2 if the value didn't get written into the register properly.
-        /// </summary>
-        public byte SetDebounceTime(ushort time)
-        {
-            return (byte)_i2cBus.WriteDoubleRegisterWithReadback(Register.BUTTON_DEBOUNCE_TIME, time);
-        }
-
-        #region Interrupt Status/Configuration
-
-        /// <summary>
-        /// When called, the interrupt will be configured to trigger when the button is pressed.
-        /// If <see cref="EnableClickedInterrupt"/> has also been called,
-        /// then the interrupt will trigger on either a push or a click.
-        /// </summary>
-        public byte EnablePressedInterrupt()
-        {
-            var interrupt =
-                new InterruptConfigBitField(_i2cBus.ReadSingleRegister(Register.INTERRUPT_CONFIG))
-                {
-                    PressedEnable = true
-                };
-            return _i2cBus.WriteSingleRegisterWithReadback(Register.INTERRUPT_CONFIG, interrupt.InterruptConfigValue);
-        }
-
-        /// <summary>
-        /// When called, the interrupt will no longer be configured to trigger when the button is pressed.
-        /// If <see cref="EnableClickedInterrupt"/> has also been called,
-        /// then the interrupt will still trigger on the button click.
-        /// </summary>
-        public byte DisablePressedInterrupt()
-        {
-            var interrupt =
-                new InterruptConfigBitField(_i2cBus.ReadSingleRegister(Register.INTERRUPT_CONFIG))
-                {
-                    PressedEnable = false
-                };
-            return _i2cBus.WriteSingleRegisterWithReadback(Register.INTERRUPT_CONFIG, interrupt.InterruptConfigValue);
-        }
-
-        /// <summary>
-        /// When called, the interrupt will be configured to trigger when the button is clicked.
-        /// If <see cref="EnablePressedInterrupt"/> has also been called,
-        /// then the interrupt will trigger on either a push or a click.
-        /// </summary>
-        public byte EnableClickedInterrupt()
-        {
-            var interrupt =
-                new InterruptConfigBitField(_i2cBus.ReadSingleRegister(Register.INTERRUPT_CONFIG))
-                {
-                    ClickedEnable = true
-                };
-
-            return _i2cBus.WriteSingleRegisterWithReadback(Register.INTERRUPT_CONFIG, interrupt.InterruptConfigValue);
-        }
-
-        /// <summary>
-        /// When called, the interrupt will no longer be configured to trigger when the button is clicked.
-        /// If <see cref="EnablePressedInterrupt"/> has also been called,
-        /// then the interrupt will still trigger on the button press.
-        /// </summary>
-        public byte DisableClickedInterrupt()
-        {
-            var interrupt =
-                new InterruptConfigBitField(_i2cBus.ReadSingleRegister(Register.INTERRUPT_CONFIG))
-                {
-                    ClickedEnable = false
-                };
-
-            return _i2cBus.WriteSingleRegisterWithReadback(Register.INTERRUPT_CONFIG, interrupt.InterruptConfigValue);
-        }
-
-        /// <summary>
-        /// Resets the interrupt configuration back to defaults.
-        /// </summary>
-        public byte ResetInterruptConfig()
-        {
-            var interrupt = new InterruptConfigBitField
-            {
-                PressedEnable = true,
-                ClickedEnable = true
-            };
-            var interruptValue = _i2cBus.WriteSingleRegisterWithReadback(Register.INTERRUPT_CONFIG, interrupt.InterruptConfigValue);
-
-            var status = new StatusRegisterBitField
-            {
-                EventAvailable = false
-            };
-            _i2cBus.WriteSingleRegisterWithReadback(Register.BUTTON_STATUS, status.StatusRegisterValue);
-
-            return interruptValue;
-        }
-
-        #endregion
-
-        #region Queue Manipulation
-
-        /// <summary>
-        /// Returns whether the queue of button press timestamps is full.
-        /// </summary>
-        public bool IsPressedQueueFull()
-        {
-            var pressedQueue = new QueueStatusBitField(_i2cBus.ReadSingleRegister(Register.PRESSED_QUEUE_STATUS));
-            return pressedQueue.IsFull;
-        }
-
-        /// <summary>
-        /// Returns whether the queue of button press timestamps is empty.
-        /// </summary>
-        public bool IsPressedQueueEmpty()
-        {
-            var pressedQueue = new QueueStatusBitField(_i2cBus.ReadSingleRegister(Register.PRESSED_QUEUE_STATUS));
-            return pressedQueue.IsEmpty;
-        }
-
-        /// <summary>
-        /// Returns how many milliseconds it has been since the last button press.
-        /// Since this returns a 32-bit unsigned int, it will roll over about every 50 days.
-        /// </summary>
-        public uint TimeSinceLastPress()
-        {
-            return _i2cBus.ReadQuadRegister(Register.PRESSED_QUEUE_FRONT);
-        }
-
-        /// <summary>
-        /// Returns how many milliseconds it has been since the first button press.
-        /// Since this returns a 32-bit unsigned int, it will roll over about every 50 days.
-        /// </summary>
-        public uint TimeSinceFirstPress()
-        {
-            return _i2cBus.ReadQuadRegister(Register.PRESSED_QUEUE_BACK);
-        }
-
-        /// <summary>
-        /// Returns the oldest value in the queue (milliseconds since first button press), and then removes it.
-        /// </summary>
-        public uint PopPressedQueue()
-        {
-            var timeSinceFirstPress = TimeSinceFirstPress(); // Take the oldest value on the queue
-
-            var pressedQueue =
-                new QueueStatusBitField(_i2cBus.ReadSingleRegister(Register.PRESSED_QUEUE_STATUS))
-                {
-                    PopRequest = true
-                };
-
-            // Remove the oldest value from the queue
-            _i2cBus.WriteSingleRegister(Register.PRESSED_QUEUE_STATUS, pressedQueue.QueueStatusValue);
-
-            return timeSinceFirstPress; // Return the value we popped
-        }
-
-        /// <summary>
-        /// Returns whether the queue of button click timestamps is full.
-        /// </summary>
-        public bool IsClickedQueueFull()
-        {
-            var clickedQueue = new QueueStatusBitField(_i2cBus.ReadSingleRegister(Register.CLICKED_QUEUE_STATUS));
-            return clickedQueue.IsFull;
-        }
-
-        /// <summary>
-        /// Returns whether the queue of button click timestamps is empty.
-        /// </summary>
-        public bool IsClickedQueueEmpty()
-        {
-            var clickedQueue = new QueueStatusBitField(_i2cBus.ReadSingleRegister(Register.CLICKED_QUEUE_STATUS));
-            return clickedQueue.IsEmpty;
-        }
-
-        /// <summary>
-        /// Returns how many milliseconds it has been since the last button click.
-        /// Since this returns a 32-bit unsigned int, it will roll over about every 50 days.
-        /// </summary>
-        public uint TimeSinceLastClick()
-        {
-            return _i2cBus.ReadQuadRegister(Register.CLICKED_QUEUE_FRONT);
-        }
-
-        /// <summary>
-        /// Returns how many milliseconds it has been since the first button click.
-        /// Since this returns a 32-bit unsigned int, it will roll over about every 50 days.
-        /// </summary>
-        public uint TimeSinceFirstClick()
-        {
-            return _i2cBus.ReadQuadRegister(Register.CLICKED_QUEUE_BACK);
-        }
-
-        /// <summary>
-        /// Returns the oldest value in the queue (milliseconds since first button click), and then removes it.
-        /// </summary>
-        public uint PopClickedQueue()
-        {
-            var timeSinceFirstClick = TimeSinceFirstClick();
-
-            var clickedQueue =
-                new QueueStatusBitField(_i2cBus.ReadSingleRegister(Register.CLICKED_QUEUE_STATUS))
-                {
-                    PopRequest = true
-                };
-            _i2cBus.WriteSingleRegister(Register.CLICKED_QUEUE_STATUS, clickedQueue.QueueStatusValue);
-
-            return timeSinceFirstClick;
-        }
-
-        #endregion
-
-        #region LED Configuration
-
-        /// <summary>
-        /// Turns the onboard LED on with the specified brightness.
-        /// <param name="brightness">LED brightness value between 0 (off) and 255 (max). Defaults to max.</param>
-        /// </summary>
-        public bool LedOn(byte brightness = 255)
-        {
-            return LedConfig(brightness, 0, 0);
-        }
-
-        /// <summary>
-        /// Turns the onboard LED off.
-        /// </summary>
-        public bool LedOff()
-        {
-            return LedConfig(0, 0, 0);
-        }
-
-        /// <summary>
-        /// Configures the onboard LED with the given max brightness, granularity, cycle time, and off time.
-        /// <param name="brightness">LED brightness value between 0 (off) and 255 (max). Default is 255.</param>
-        /// <param name="cycleTime">Total pulse cycle time (in ms). Does not include off time.</param>
-        /// <param name="offTime">Off time between pulses (in ms). Default is 500 ms.</param>
-        /// <param name="granularity">Amount of steps it takes to get to the set brightness level (1 is fine for most applications).</param>
-        /// </summary>
-        public bool LedConfig(byte brightness, ushort cycleTime, ushort offTime, byte granularity = 1)
-        {
-            bool success = _i2cBus.WriteSingleRegister(Register.LED_BRIGHTNESS, brightness);
-            success &= _i2cBus.WriteSingleRegister(Register.LED_PULSE_GRANULARITY, granularity);
-            success &= _i2cBus.WriteDoubleRegister(Register.LED_PULSE_CYCLE_TIME, cycleTime);
-            success &= _i2cBus.WriteDoubleRegister(Register.LED_PULSE_OFF_TIME, offTime);
-            return success;
-        }
-
-        #endregion
 
         /// <inheritdoc />
         public void Dispose()
